classtricks: Refactor get_resolved_dict() to be more general, add a mode that does what you expect for classes.
The issue is that a class can also be looked at as an instance.
So suppose we have a metaclass M, a class A of type M, a class B inheriting from A, and an instance b of type B.
Then a user wanting to see a resolved dict of b would expect the contents of b, B and A.
Similarly, a user wanting a resolved dict of B would expect B, A and M.
But if the user wanted to get the contents of B *that b would inherit*, that would not include M.
This is what most people generally expect when talking about classes. We provide a cls=True option to enforce
that behaviour.