ucd: Tricks to dramatically reduce memory usage
* __slots__ for CodePoint
	Helps a little
* Instead of holding attrs dicts, hold a 'packed' dict using msgpack
	Helps a lot
* Also include aliases in that dict, so it's packed too
	Helps very little
* Eliminate most-common values in packed form by consulting a hard-coded list during pack/unpack
	Helps the most

However, still have problems with super high high-water mark of memory usage during load,
so we save a 'cache file' with the msgpacked values once, then load from that instead.
Not only does this result in an orders of mangitude smaller high-water mark, but is also much faster!

In summary, we reduce memory usage from several GB to ~100-150MB.