Rewrite stacksampler to support top-N queries
Because recording _all_ stacks is too large in many cases,
and you don't actually care about most rare stacks, this allows you to
only ask for the most common stacks (including all children, so you don't have
a child stack without a parent, and any omitted children are automatically attributed
to the last included parent).

Because the new structure is considerably longer and more involved,
we keep the original version around because being able to copy-paste a few tens of hacky lines
all contained in one function definition is surprisingly useful when debugging.